# Remove superfluous initial iteration of repeating parts of optimizer sequence