--- a/test/libsolidity/semanticTests/array/copying/array_elements_to_mapping.sol
+++ b/test/libsolidity/semanticTests/array/copying/array_elements_to_mapping.sol
@@ -52,7 +52,7 @@ contract C {
 }
 // ----
 // from_storage() -> 0x20, 2, 0x40, 0xa0, 2, 10, 11, 3, 12, 13, 14
-// gas irOptimized: 150004
+// gas irOptimized: 150022
 // gas legacy: 150745
 // gas legacyOptimized: 148678
 // from_storage_ptr() -> 0x20, 2, 0x40, 0xa0, 2, 10, 11, 3, 12, 13, 14