--- a/test/libsolidity/semanticTests/array/copying/array_to_mapping.sol
+++ b/test/libsolidity/semanticTests/array/copying/array_to_mapping.sol
@@ -37,7 +37,7 @@ contract C {
 }
 // ----
 // from_storage() -> 0x20, 2, 0x40, 0xa0, 2, 10, 11, 3, 12, 13, 14
-// gas irOptimized: 147871
+// gas irOptimized: 147889
 // gas legacy: 148896
 // gas legacyOptimized: 146901
 // from_storage_ptr() -> 0x20, 2, 0x40, 0xa0, 2, 10, 11, 3, 12, 13, 14