--- a/test/libsolidity/semanticTests/functionCall/gas_and_value_basic.sol
+++ b/test/libsolidity/semanticTests/functionCall/gas_and_value_basic.sol
@@ -38,8 +38,8 @@ contract test {
 }
 // ----
 // constructor(), 20 wei ->
-// gas irOptimized: 120226
-// gas irOptimized code: 132400
+// gas irOptimized: 120912
+// gas irOptimized code: 140000
 // gas legacy: 130568
 // gas legacy code: 261000
 // gas legacyOptimized: 121069