--- a/test/libsolidity/semanticTests/types/mapping/copy_from_mapping_to_mapping.sol
+++ b/test/libsolidity/semanticTests/types/mapping/copy_from_mapping_to_mapping.sol
@@ -29,6 +29,6 @@ contract C {
 }
 // ----
 // f() -> 0x20, 7, 8, 9, 0xa0, 13, 2, 0x40, 0xa0, 2, 3, 4, 2, 3, 4
-// gas irOptimized: 197082
+// gas irOptimized: 197116
 // gas legacy: 199891
 // gas legacyOptimized: 196817