--- a/test/libyul/yulOptimizerTests/fullSuite/unusedFunctionParameterPruner_loop.yul
+++ b/test/libyul/yulOptimizerTests/fullSuite/unusedFunctionParameterPruner_loop.yul
@@ -20,18 +20,46 @@
 // {
 //     {
 //         f()
-//         f()
-//         f()
+//         f_72()
+//         f_73()
 //         sstore(0, 1)
 //     }
 //     function f()
 //     {
+//         let a := 1
 //         let b := 10
-//         let a := calldataload(0)
-//         let _1 := iszero(a)
+//         let a_1 := calldataload(0)
+//         let _1 := iszero(a_1)
 //         for { } iszero(b) { b := add(b, not(0)) }
 //         {
-//             mstore(a, 0)
+//             a := a_1
+//             mstore(a_1, 0)
+//             if _1 { leave }
+//         }
+//     }
+//     function f_72()
+//     {
+//         let a := 2
+//         let b := 10
+//         let a_1 := calldataload(0)
+//         let _1 := iszero(a_1)
+//         for { } iszero(b) { b := add(b, not(0)) }
+//         {
+//             a := a_1
+//             mstore(a_1, 0)
+//             if _1 { leave }
+//         }
+//     }
+//     function f_73()
+//     {
+//         let a := 3
+//         let b := 10
+//         let a_1 := calldataload(0)
+//         let _1 := iszero(a_1)
+//         for { } iszero(b) { b := add(b, not(0)) }
+//         {
+//             a := a_1
+//             mstore(a_1, 0)
 //             if _1 { leave }
 //         }
 //     }