… test/libsolidity/semanticTests/array/copying/elements_of_nested_array_of_structs_calldata_to_storage.sol
… test/libsolidity/semanticTests/array/copying/elements_of_nested_array_of_structs_memory_to_storage.sol
… test/libsolidity/semanticTests/array/copying/nested_array_of_structs_calldata_to_storage.sol
… test/libsolidity/semanticTests/array/copying/nested_array_of_structs_memory_to_storage.sol
… test/libsolidity/semanticTests/externalContracts/deposit_contract.sol
… test/libsolidity/semanticTests/structs/copy_substructures_from_mapping.sol
… test/libsolidity/semanticTests/structs/copy_substructures_to_mapping.sol
… test/libsolidity/semanticTests/structs/copy_to_mapping.sol