The optimizer will repeat each bracketed segment of the optimizer step sequence  at least twice, because the cost is only measured after the first iteration. This results in unnecessary repetitions in some cases.

Hard to tell how much it actually affects the optimization time in practice. From my observations, a single iteration is often insufficient so many contracts would not be affected. On the other hand, for the lucky ones where one iteration is enough, it could cut optimization time almost in half. I need to see CI timings to say more.

Looks like I unwittingly introduced this regression back in #12102. Originally this loop looked like this and I moved the check to the end without updating the initialization: https://github.com/ethereum/solidity/blob/f35694f655e81d503b3cbc1aee324ea4cc848a3c/libyul/optimiser/Suite.cpp#L84-L92