@cameel 2024-02-14
Looks like the actual differences are minimal, if any, and go in both directions. For most jobs they are within the margin of error.

| Job                                  | [`develop`](https://app.circleci.com/pipelines/github/ethereum/solidity/32859/workflows/dc2abfa2-1a62-4f0c-828c-3e63830f7525) | [this PR](https://app.circleci.com/pipelines/github/ethereum/solidity/32881/workflows/83d22b3d-4d8b-48c2-b4a4-57b8d42e6136) |
|--------------------------------------|---------|---------|
| `b_bytecode_ubu-via-ir-optimize`     |  9m 39s | 10m 12s |
| `b_bytecode_osx-via-ir-optimize`     |  4m 23s |  4m 18s |
| `t_native_test_ext_zeppelin`         | 12m 17s | 10m 49s |
| `t_native_test_ext_brink`            |  1m 21s |  1m 40s |
| `t_native_test_ext_elementfi`        |  4m 41s |  4m 16s |
| `t_native_test_ext_ens`              |  2m 25s |  2m 41s |
| `t_native_test_ext_perpetual_pools`  | 18m 52s | 19m 16s |
| `t_native_test_ext_prb_math`         |  4m 10s |  4m  9s |
| `t_native_test_ext_uniswap`          |  9m 35s |  9m 22s |
| `t_native_test_ext_yield_liquidator` |  2m 26s |  2m 36s |

The only one that stands out is the OpenZeppelin external test, where it's a 12% reduction, though, this could still be just CI variance.