@cameel 2024-04-22
### Gas diff stats
| File name                                                                   | IR optimized   | Legacy optimized   | Legacy   |
|-----------------------------------------------------------------------------|----------------|--------------------|----------|
| `functionCall/gas_and_value_brace_syntax.sol`                               | 3%             |                    |          |
| `functionCall/gas_and_value_basic.sol`                                      | 3%             |                    |          |
| `externalContracts/deposit_contract.sol`                                    | +0%            |                    |          |
| `libraries/internal_types_in_library.sol`                                   | +0%            |                    |          |
| `userDefinedValueType/calldata.sol`                                         | +0%            |                    |          |
| `types/mapping/copy_from_mapping_to_mapping.sol`                            | +0%            |                    |          |
| `array/copying/array_to_mapping.sol`                                        | +0%            |                    |          |
| `array/copying/array_elements_to_mapping.sol`                               | +0%            |                    |          |
| `array/copying/nested_array_of_structs_memory_to_storage.sol`               | +0%            |                    |          |
| `array/copying/elements_of_nested_array_of_structs_memory_to_storage.sol`   | +0%            |                    |          |
| `various/skip_dynamic_types_for_structs.sol`                                | +0%            |                    |          |
| `externalContracts/snark.sol`                                               | +0%            |                    |          |
| `array/copying/nested_array_of_structs_calldata_to_storage.sol`             | +0%            |                    |          |
| `array/copying/elements_of_nested_array_of_structs_calldata_to_storage.sol` | +0%            |                    |          |
| `array/copying/storage_memory_packed_dyn.sol`                               | -0%            |                    |          |
| `structs/copy_to_mapping.sol`                                               | -0%            |                    |          |
| `structs/copy_substructures_from_mapping.sol`                               | -0%            |                    |          |
| `structs/copy_substructures_to_mapping.sol`                                 | -0%            |                    |          |
| `array/copying/calldata_array_to_mapping.sol`                               | -0%            |                    |          |