@cameel 2024-04-22
### External tests (`ir-optimize-evm+yul`)
|          project |  bytecode_size | deployment_gas |     method_gas |
|:----------------:|---------------:|---------------:|---------------:|
|            brink |           `0%` |                |                |
|           colony |           `0%` |                |                |
|        elementfi | **`+0.21% ❌`** |                |                |
|              ens |  **`-0.5% ✅`** |          `-0%` |           `0%` |
|            euler | **`+0.01% ❌`** |                |                |
|           gnosis |                |                |                |
|              gp2 |           `0%` |                |                |
|  perpetual-pools |          `+0%` |          `+0%` | **`-0.02% ✅`** |
|    pool-together | **`-0.14% ✅`** |                |                |
|          uniswap | **`-0.01% ✅`** |                |                |
| yield_liquidator | **`+0.01% ❌`** | **`+0.01% ❌`** |          `+0%` |
|         zeppelin | **`-0.01% ✅`** | **`-0.04% ✅`** | **`-0.06% ✅`** |

#### Timing
| Project          |   Before |   After |   Diff |
|:-----------------|---------:|--------:|-------:|
| brink            |      5 s |     5 s |    0 s |
| colony           |    105 s |   117 s |   12 s |
| elementfi        |    172 s |   157 s |  -15 s |
| ens              |     40 s |    39 s |   -1 s |
| euler            |     52 s |    60 s |    8 s |
| gnosis           |          |         |        |
| gp2              |     42 s |    46 s |    4 s |
| perpetual-pools  |     81 s |    77 s |   -4 s |
| pool-together    |     56 s |    46 s |  -10 s |
| uniswap          |     82 s |    73 s |   -9 s |
| yield_liquidator |     26 s |    38 s |   12 s |
| zeppelin         |    261 s |   245 s |  -16 s |