@cameel 2024-04-22
### Quick and dirty non-CI benchmark

Since CI timing is almost useless here due to a huge variation, I tried to benchmark it manually using a few projects that can be compiled right away via foundry.

<details><summary>script</summary>

```bash
function benchmark_foundry_project {
    local subdir="$1"
    local url="$2"
    local solc_path="$3"

    git clone "$url" "$subdir" --depth=1
    pushd "$subdir"
    forge install
    /usr/bin/time --output ../timing.json --append --format "{\"$subdir\": {\"real\": %e, \"user\": %U, \"sys\": %S}}" \
        forge build --use "$solc_path" --optimize --via-ir --evm-version cancun --offline --no-cache
    popd
    rm -rf "./${subdir}"
}

rm timing.json

solc_binary=/tmp/solc-develop
benchmark_foundry_project uniswap-before https://github.com/Uniswap/v4-core "$solc_binary"
benchmark_foundry_project seaport-before https://github.com/ProjectOpenSea/seaport-core "$solc_binary"

solc_binary=/tmp/solc-fix-superfluous-iterations-in-optimizer-sequence
benchmark_foundry_project uniswap-after https://github.com/Uniswap/v4-core "$solc_binary"
benchmark_foundry_project seaport-after https://github.com/ProjectOpenSea/seaport-core "$solc_binary"

cat timing.json | jq --slurp add
```

</details>

I also wanted to use OpenZeppelin but it sadly fails with "Stack Too Deep" for both binaries (despite the memory guard being present). Same with a 0.8.25 release binary.

#### Results

| Project | run |   before   |   after    |
|---------|-----|------------|------------|
| uniswap |   1 |    308 s   |    280 s   |
| uniswap |   2 |    296 s   |    279 s   |
| uniswap |   3 |    289 s   |    277 s   |
| uniswap | avg |  **298 s** |  **278 s** |
| seaport |   1 |     25 s   |     17 s   |
| seaport |   2 |     19 s   |     17 s   |
| seaport |   3 |     19 s   |     23 s   |
| seaport | avg |   **21 s** |   **19 s** |

Looks like this has a lot of variance too, but there's a clear and reproducible speed improvement on the order of 10%.